refactor: collect process to use behaviors
