+++ packages/cli/src/lib/core/utils/processing/behaviors.ts
@@ -2,13 +2,13 @@ import { RcJson } from '@push-based/user-flow';
 import { CLIProcess } from './types';
 
 export function run(
-  tasks: CLIProcess[],
+  tasks: CLIProcess[]
 ): CLIProcess {
   return (config) => concat(tasks)(config);
 }
 
 export function concat(processes: CLIProcess[]): CLIProcess {
-  return async function (r: RcJson): Promise<RcJson> {
+  return async function(r: RcJson): Promise<RcJson> {
     return await processes.reduce(
       async (cfg, processor) => await processor(await cfg),
       Promise.resolve(r)
@@ -16,8 +16,20 @@ export function concat(processes: CLIProcess[]): CLIProcess {
   };
 }
 
+
+export function ifThenElse(condition: (r: RcJson) => boolean, thenProcess: CLIProcess, elseProcess?: CLIProcess): CLIProcess {
+  return async function(r: RcJson): Promise<RcJson> {
+    const conditionResult = await condition(r);
+    if (conditionResult) {
+      return thenProcess(r);
+    } else {
+      return elseProcess ? elseProcess(r) : Promise.resolve(r);
+    }
+  };
+}
+
 export function tap(process: CLIProcess): CLIProcess {
-  return async function (d: RcJson): Promise<RcJson> {
+  return async function(d: RcJson): Promise<RcJson> {
     await process(d);
     return Promise.resolve(d);
   };
